refactor: Reorganize message data classes for improved structure and clarity
- Moved all WebSocket-related messages to new package: data.messages
- Introduced GameMessage interface for consistent message structure
