--- app/src/main/java/at/aau/serg/websocketbrokerdemo/data/ChatMessage.kt
@@ -1,7 +1,0 @@
-package at.aau.serg.websocketbrokerdemo.data
-
-data class ChatMessage(
-    val type:String="CHAT_MESSAGE",
-    val playerId:String,
-    val message:String
-)
