--- app/src/main/java/at/aau/serg/websocketbrokerdemo/data/DiceRollMessage.kt
@@ -1,7 +1,0 @@
-package at.aau.serg.websocketbrokerdemo.data
-
-data class DiceRollMessage(
-    val type:     String,
-    val playerId: String,
-    val value:    Int
-)
