+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/ChatMessage.kt
@@ -1,0 +1,7 @@
+package at.aau.serg.websocketbrokerdemo.data.messages
+
+data class ChatMessage(
+   override val type:String="CHAT_MESSAGE",
+    val playerId:String,
+    val message:String
+):GameMessage
