+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/DiceRollMessage.kt
@@ -1,0 +1,7 @@
+package at.aau.serg.websocketbrokerdemo.data.messages
+
+data class DiceRollMessage(
+   override val type:String,
+    val playerId: String,
+    val value:  Int
+):GameMessage
