+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/DrawnCardMessage.kt
@@ -1,10 +1,10 @@
-package at.aau.serg.websocketbrokerdemo.data
+package at.aau.serg.websocketbrokerdemo.data.messages
 
 import com.google.gson.JsonObject
 
 data class DrawnCardMessage(
-    val type: String,       // == "CARD_DRAWN"
+   override val type: String,       // == "CARD_DRAWN"
     val playerId: String,
     val cardType: String,   // "CHANCE" or "COMMUNITY_CHEST"
     val card: JsonObject    // <— raw JSON object
-)
+):GameMessage
