+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/GameMessage.kt
@@ -1,0 +1,5 @@
+package at.aau.serg.websocketbrokerdemo.data.messages
+
+interface GameMessage {
+    val type:String
+}
